Adds 'brightpixel' modulation type.
The behavrior and usage is the same as a 'square'
amplitude modulation. Even the command-line parameters
are the same.

However, there is now some code that makes assumptions
about the size of the simultation! In particular, the 'brightpixel'
type restricts the phase of the square wave amplitude modulation
such that the ON portion of the square wave will not straddle
the beginning or ending of the simulation -- it won't wrap
around from the end to the beginning of the simulation.
This prevents the first 2% and last 2% of a simulation from containing
the bright pixel, so long as the bright pixel length is less
than 2% of the total length of the simulation signal.

Updates simulation version number.